Trait eyeball_im_util::vector::VectorDiffContainer

pub trait VectorDiffContainer: VectorDiffContainerOps<Self::Element, Family = Self::Family> {
    type Element: Clone + Send + Sync + 'static;
}

Abstraction over stream items that the adapters in this module can deal with.

Required Associated Types

type Element: Clone + Send + Sync + 'static

The element type of the Vector that diffs are being handled for.

Object Safety

This trait is not object safe.