Trait eyeball_im_util::vector::VectorObserverExt

pub trait VectorObserverExt<T>: VectorObserver<T>
where
    T: Clone + Send + Sync + 'static,
    <Self::Stream as Stream>::Item: VectorDiffContainer<Element = T>,
{
    // Provided methods
    fn filter<F>(self, f: F) -> (Vector<T>, Filter<Self::Stream, F>)
       where F: Fn(&T) -> bool { ... }
    fn filter_map<U, F>(self, f: F) -> (Vector<U>, FilterMap<Self::Stream, F>)
       where U: Clone,
             F: Fn(T) -> Option<U> { ... }
    fn limit(
        self,
        limit: usize
    ) -> (Vector<T>, Limit<Self::Stream, EmptyLimitStream>) { ... }
    fn dynamic_limit<L>(self, limit_stream: L) -> Limit<Self::Stream, L>
       where L: Stream<Item = usize> { ... }
    fn dynamic_limit_with_initial_value<L>(
        self,
        initial_limit: usize,
        limit_stream: L
    ) -> (Vector<T>, Limit<Self::Stream, L>)
       where L: Stream<Item = usize> { ... }
}

Convenience methods for VectorObservers.

See that trait for which types implement this.

Provided Methods

fn filter<F>(self, f: F) -> (Vector<T>, Filter<Self::Stream, F>)
where F: Fn(&T) -> bool,

Filter the vector’s values with the given function.

fn filter_map<U, F>(self, f: F) -> (Vector<U>, FilterMap<Self::Stream, F>)
where U: Clone, F: Fn(T) -> Option<U>,

Filter and map the vector’s values with the given function.

fn limit( self, limit: usize ) -> (Vector<T>, Limit<Self::Stream, EmptyLimitStream>)

Limit the observed values to limit.

See Limit for more details.

fn dynamic_limit<L>(self, limit_stream: L) -> Limit<Self::Stream, L>
where L: Stream<Item = usize>,

Limit the observed values to a number of items determined by the given stream.

See Limit for more details.

fn dynamic_limit_with_initial_value<L>( self, initial_limit: usize, limit_stream: L ) -> (Vector<T>, Limit<Self::Stream, L>)
where L: Stream<Item = usize>,

Limit the observed values to initial_limit items initially, and update the limit with the value from the given stream.

See Limit for more details.

Object Safety

This trait is not object safe.